requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
pytest==7.4.0
